feat: add export expense functionality
Adds an export button to the expense list, allowing users to export
their expenses as a JSON file. This is implemented using a new
JavaScript interface in the Android app to handle file creation.

This feature provides users with a way to back up their expense data and
potentially use it with other applications.